requirements.txt
impacket
pycryptodome
